Rewrite _contains() to expect array as string of elements
This follows the philosophy of everything being strings and should
eliminate some of the ambiguity around how these elements are being
handled. By making _contains() generate the array on assignment using
all of the arguments following the first one, it should be clearer that
the second argument is a list of array elements.

Also, as a part of this process, use underscore prefixed variables
rather than the local keyword. local isn't used elsewhere in the
program, so using it here is inconsistent. If local is going to be used,
it should be used everywhere.

These updates also appear to eliminate array issues with Bash 3.2, which
is the default on OS X.