Have the -h and --help flags always invoke the help command.
In versions prior to this change, the -h and --help flags had to appear
before --version or an non-option argument in order to set the command
to 'help'. As a result, `$_ME example -h` would pass the -h option to
the command rather than behave like `$_ME help example`, as might be
expected in some circumstances.

Using a multi-step assignment processes, the $cmd variable can be set to
'help' no matter where it appears in the arguments, and any instances of
the --version flag can indicate that this invocation of the program
would prefer to use the 'version' command if no other command is found.

In the second step, the command is set to the argument in the first
position if the help command did not already set it. Finally, if the
command has still not be set, the program checks to see if 'version' is
expected and, if so, sets the command to that.

The downside to this approach is that the -h, --help, and --version
flags can no longer be used as arguments within the commands
themselves, which might be desired if the commands are wrapping other
commands that might expect that option.