Simplify option `while` test and fix option case in bash-simple-plus.
The previous `while` test resulted in behavior that was slightly
difficult to follow, whereas this test from bash-commands works as
expected and is easier to understand.

The unexpected option fallback case neglected to have a leading '-', an
issue that was apparently masked by the previous `while` test, so add a
leading '-' to produce the expected behavior.