Use braces in all variable references (continued).
This is a continuation of 3248061 that converts more variable references
to the braces format.

Braces are only required in certain cases, but the cognitive overhead in
keeping track of which cases require braces can be reduced by simply
always using them.

Example: `${NAME}`

Retain more widely-used braceless `$NAME` convention in documentation.